It's in the branch name
Nothing much to say here; changed all of a few lines. Floats are now doubles, int is now size_t.

+++ perftest/normal_static.cpp
@@ -16,16 +16,16 @@
 using namespace dense::stochastic;
 
 int main() {
-  std::normal_distribution<float> d(5,2); 
+  std::normal_distribution<double> d(5,2);
   std::default_random_engine generator;
-  std::vector<float> weights = {};
+  std::vector<double> weights = {};
   int sum = 1;
   
   for(int i = 0; i < WEIGHTNUM; i++){
     weights.push_back(d(generator));
   }	      
 
-  float minweight = *std::min_element(weights.begin(), weights.end());
+  double minweight = *std::min_element(weights.begin(), weights.end());
   for(int i = 0; i < WEIGHTNUM; i++){
     weights[i] -= minweight;
   }
